refactor(solver): split interner core into storage and display submodules (#13000)
crates/tsz-solver/src/intern/core/interner.rs exceeded the repo's 2000-line
hard limit (2007 lines) and combined three distinct concerns in one file:
sharded storage layout, intern/lookup orchestration, and diagnostic display
provenance.

Extract two cohesive submodules under interner/, moving code verbatim:
- storage.rs: sharded TypeData storage (TypeShard/TypeShardInner) and the
  concurrent slice/value component interners plus CachedUnionMember -- pure
  data layout with no knowledge of TypeInterner semantics.
- display.rs: display/provenance side tables (fresh-literal properties,
  evaluated-application alias names, conditional-alias bases, as-written union
  member origin) and the union_origin_overrides_canonical_* predicates.

interner.rs keeps the TypeInterner struct, the intern/lookup hot paths, and
the component accessors (now 1249 lines). Storage items are scoped to
pub(in crate::intern::core) to match their actual consumers (parent interner
module + sibling constructors module) rather than leaking crate-wide. No
public API, type identity, or behavior changes; #[inline] preserved on all
hot-path methods.
