+++ crates/tsz-solver/src/intern/core/interner.rs
@@ -1,8 +1,12 @@
 //! Core implementation of the type interning engine.
 //!
-//! This module contains all data structures and methods for the `TypeInterner`,
-//! including sharded storage, concurrent slice/value interners, and type
-//! construction convenience methods.
+//! This module owns the `TypeInterner` struct, its intern/lookup hot paths, and
+//! the component accessors. Concern-specific pieces live in submodules:
+//! - `storage`: sharded `TypeData` storage and the slice/value component interners
+//!   (pure data layout).
+//! - `display`: diagnostic display provenance (fresh-literal properties, alias
+//!   names, union member origin).
+//! - `cache`: the thread-local intern/lookup fast-path cache.
 
 use crate::def::DefId;
 use crate::types::{
@@ -19,12 +23,17 @@ use rustc_hash::{FxBuildHasher, FxHashMap, FxHasher};
 use smallvec::SmallVec;
 use std::hash::{Hash, Hasher};
 use std::sync::{
-    Arc, OnceLock, RwLock,
+    Arc, OnceLock,
     atomic::{AtomicBool, AtomicU32, Ordering},
 };
 use tsz_common::interner::{Atom, ShardedInterner};
 
 mod cache;
+mod display;
+mod storage;
+
+pub(super) use storage::{CachedUnionMember, TypeShard};
+use storage::{ConcurrentSliceInterner, ConcurrentValueInterner};
 
 /// Global counter for assigning unique `instance_id`s to `TypeInterner`
 /// instances. `0` is reserved as "empty/no-interner" so it will never match
@@ -97,273 +106,6 @@ pub(crate) struct InternedTypeLimitContext {
     pub(crate) current_count: usize,
     pub(crate) max_interned_types: usize,
     pub(crate) fallback_type: TypeId,
-}
-
-/// Cached data for a union member, pre-fetched to avoid redundant DashMap/arena
-/// lookups during sort comparisons. Each field corresponds to a lookup that
-/// `compare_union_members` would otherwise perform per comparison.
-pub(super) struct CachedUnionMember {
-    /// The original TypeId
-    pub(super) id: TypeId,
-    /// Result of `builtin_sort_key(id)` - `Some` for intrinsic/builtin types
-    pub(super) builtin_key: Option<u32>,
-    /// Result of `self.lookup(id)` - the TypeData for non-builtin types
-    pub(super) data: Option<TypeData>,
-    /// For Object/ObjectWithIndex: the symbol's raw u32 (if the shape has a symbol)
-    pub(super) obj_symbol: Option<u32>,
-    /// For anonymous Object/ObjectWithIndex: the `ShapeId`'s raw u32
-    pub(super) obj_anon_shape: Option<u32>,
-    /// For Callable: the symbol's raw u32 (if the shape has a symbol)
-    pub(super) callable_symbol: Option<u32>,
-    /// Monotonic allocation counter for source-order sorting
-    pub(super) alloc_order: Option<u32>,
-}
-
-/// Inner data for a `TypeShard`, lazily initialized.
-pub(super) struct TypeShardInner {
-    /// Map from `TypeData` to local index within this shard
-    key_to_index: DashMap<TypeData, u32, FxBuildHasher>,
-    /// Flat array from local index to `TypeData`.
-    /// Sequential indices make a Vec far faster than `DashMap` for reverse lookup.
-    /// Protected by `RwLock`: reads are uncontended in single-threaded use (~1 cycle),
-    /// writes only happen during intern (append-only).
-    index_to_key: RwLock<Vec<TypeData>>,
-    /// Per-shard allocation order (parallel to `index_to_key`).
-    /// Stores the global monotonic order counter at time of interning.
-    alloc_order: RwLock<Vec<u32>>,
-}
-
-/// A single shard of the type interned storage.
-///
-/// Uses `OnceLock` for lazy initialization - `DashMaps` are only allocated
-/// when the shard is first accessed, reducing startup overhead.
-pub(super) struct TypeShard {
-    /// Lazily initialized inner maps
-    pub(super) inner: OnceLock<TypeShardInner>,
-    /// Atomic counter for allocating new indices in this shard
-    /// Kept outside `OnceLock` for fast checks without initialization
-    pub(super) next_index: AtomicU32,
-}
-
-impl TypeShard {
-    const fn new() -> Self {
-        Self {
-            inner: OnceLock::new(),
-            next_index: AtomicU32::new(0),
-        }
-    }
-
-    /// Get the inner maps, initializing on first access
-    #[inline]
-    fn get_inner(&self) -> &TypeShardInner {
-        self.inner.get_or_init(|| TypeShardInner {
-            key_to_index: DashMap::with_hasher(FxBuildHasher),
-            index_to_key: RwLock::new(Vec::with_capacity(256)),
-            alloc_order: RwLock::new(Vec::with_capacity(256)),
-        })
-    }
-
-    /// Check if a key exists without initializing the shard
-    #[inline]
-    fn is_empty(&self) -> bool {
-        self.next_index.load(Ordering::Relaxed) == 0
-    }
-}
-
-/// Inner data for `ConcurrentSliceInterner`, lazily initialized.
-pub(super) struct SliceInternerInner<T> {
-    /// Flat array from ID to slice value. Sequential IDs make Vec optimal for reverse lookup.
-    items: RwLock<Vec<Arc<[T]>>>,
-    map: DashMap<Arc<[T]>, u32, FxBuildHasher>,
-}
-
-/// Slice interner using flat Vec for reverse lookup.
-/// Uses lazy initialization to defer allocation until first use.
-pub(super) struct ConcurrentSliceInterner<T> {
-    pub(super) inner: OnceLock<SliceInternerInner<T>>,
-    pub(super) next_id: AtomicU32,
-}
-
-impl<T> ConcurrentSliceInterner<T>
-where
-    T: Eq + Hash + Clone + Send + Sync + 'static,
-{
-    const fn new() -> Self {
-        Self {
-            inner: OnceLock::new(),
-            next_id: AtomicU32::new(1), // Reserve 0 for empty
-        }
-    }
-
-    #[inline]
-    fn get_inner(&self) -> &SliceInternerInner<T> {
-        self.inner.get_or_init(|| {
-            let empty: Arc<[T]> = Arc::from(Vec::new());
-            let mut items_vec = Vec::with_capacity(256);
-            items_vec.push(Arc::clone(&empty)); // id 0 = empty
-            let map = DashMap::with_hasher(FxBuildHasher);
-            map.insert(empty, 0);
-            SliceInternerInner {
-                items: RwLock::new(items_vec),
-                map,
-            }
-        })
-    }
-
-    #[inline]
-    fn intern(&self, items_slice: &[T]) -> u32 {
-        if items_slice.is_empty() {
-            return 0;
-        }
-
-        let inner = self.get_inner();
-
-        // PERF: Try lookup with borrowed slice first to avoid Vec+Arc allocation on cache hits.
-        // Arc<[T]>: Borrow<[T]> enables DashMap lookup with &[T] key.
-        if let Some(ref_entry) = inner.map.get(items_slice) {
-            return *ref_entry.value();
-        }
-
-        // Cache miss -- allocate for insertion
-        let temp_arc: Arc<[T]> = Arc::from(items_slice.to_vec());
-
-        // Allocate new ID
-        let id = self.next_id.fetch_add(1, Ordering::Relaxed);
-
-        // Double-check: another thread might have inserted while we allocated
-        match inner.map.entry(std::sync::Arc::clone(&temp_arc)) {
-            dashmap::mapref::entry::Entry::Vacant(e) => {
-                e.insert(id);
-                {
-                    // T2.4 instrumentation: wrap the write-lock acquisition
-                    // so contention on the slice-interner's `items` vec lands
-                    // in the lock-wait histogram alongside the per-shard
-                    // TypeData writes. With `perf-counters-timing` OFF this
-                    // wrapper compiles to a direct closure call.
-                    let mut vec = tsz_common::perf_counters::time_shard_write(0, || {
-                        inner.items.write().expect("interner items lock poisoned")
-                    });
-                    while vec.len() < id as usize {
-                        vec.push(Arc::clone(&temp_arc));
-                    }
-                    vec.push(temp_arc);
-                }
-                id
-            }
-            dashmap::mapref::entry::Entry::Occupied(e) => *e.get(),
-        }
-    }
-
-    #[inline]
-    fn get(&self, id: u32) -> Option<Arc<[T]>> {
-        // For id 0, return from the initialized inner (which has the pre-allocated
-        // empty Arc) instead of creating a new Arc::from(Vec::new()) on every call.
-        let inner = if id == 0 {
-            self.get_inner()
-        } else {
-            self.inner.get()?
-        };
-        let vec = inner.items.read().ok()?;
-        vec.get(id as usize).cloned()
-    }
-
-    #[inline]
-    fn empty(&self) -> Arc<[T]> {
-        let inner = self.get_inner();
-        let vec = inner.items.read().expect("interner items lock poisoned");
-        vec.first()
-            .cloned()
-            .unwrap_or_else(|| Arc::from(Vec::new()))
-    }
-}
-
-/// Inner data for `ConcurrentValueInterner`, lazily initialized.
-pub(super) struct ValueInternerInner<T> {
-    /// Flat array from ID to value. Sequential IDs make Vec optimal for reverse lookup.
-    items: RwLock<Vec<Arc<T>>>,
-    map: DashMap<Arc<T>, u32, FxBuildHasher>,
-}
-
-/// Value interner using flat Vec for reverse lookup.
-/// Uses lazy initialization to defer allocation until first use.
-pub(super) struct ConcurrentValueInterner<T> {
-    pub(super) inner: OnceLock<ValueInternerInner<T>>,
-    pub(super) next_id: AtomicU32,
-}
-
-impl<T> ConcurrentValueInterner<T>
-where
-    T: Eq + Hash + Clone + Send + Sync + 'static,
-{
-    const fn new() -> Self {
-        Self {
-            inner: OnceLock::new(),
-            next_id: AtomicU32::new(0),
-        }
-    }
-
-    #[inline]
-    fn get_inner(&self) -> &ValueInternerInner<T> {
-        self.inner.get_or_init(|| ValueInternerInner {
-            items: RwLock::new(Vec::with_capacity(128)),
-            map: DashMap::with_hasher(FxBuildHasher),
-        })
-    }
-
-    #[inline]
-    fn intern(&self, value: T) -> u32 {
-        let inner = self.get_inner();
-
-        // PERF: Try lookup with borrowed value first to avoid Arc allocation on cache hits.
-        // Most intern calls are for already-interned values, so this saves an Arc::new()
-        // (heap allocation + atomic ref count) on the hot path.
-        if let Some(ref_entry) = inner.map.get(&value) {
-            return *ref_entry.value();
-        }
-
-        // Cache miss -- allocate Arc for insertion
-        let value_arc = Arc::new(value);
-
-        // Allocate new ID
-        let id = self.next_id.fetch_add(1, Ordering::Relaxed);
-
-        // Double-check: another thread might have inserted while we allocated
-        match inner.map.entry(std::sync::Arc::clone(&value_arc)) {
-            Entry::Vacant(e) => {
-                e.insert(id);
-                {
-                    // T2.4 instrumentation: see the matching wrapper in
-                    // `ConcurrentSliceInterner::intern`. Same rationale,
-                    // same zero-cost-when-feature-off contract.
-                    let mut vec = tsz_common::perf_counters::time_shard_write(0, || {
-                        inner.items.write().expect("interner items lock poisoned")
-                    });
-                    while vec.len() < id as usize {
-                        vec.push(Arc::clone(&value_arc));
-                    }
-                    vec.push(value_arc);
-                }
-                id
-            }
-            Entry::Occupied(e) => *e.get(),
-        }
-    }
-
-    #[inline]
-    fn get(&self, id: u32) -> Option<Arc<T>> {
-        let vec = self.inner.get()?.items.read().ok()?;
-        vec.get(id as usize).cloned()
-    }
-
-    /// Get value by copy for Copy types, avoiding Arc clone overhead.
-    #[inline]
-    fn get_copy(&self, id: u32) -> Option<T>
-    where
-        T: Copy,
-    {
-        let vec = self.inner.get()?.items.read().ok()?;
-        vec.get(id as usize).map(|arc| **arc)
-    }
 }
 
 /// Type interning table with lock-free concurrent access.
@@ -1339,506 +1081,6 @@ impl TypeInterner {
     pub fn intern_object_shape(&self, shape: ObjectShape) -> ObjectShapeId {
         tsz_common::perf_counters::record_interner_object_shape_intern_call();
         ObjectShapeId(self.object_shapes.intern(shape))
-    }
-
-    /// Store display-only properties for a fresh object literal.
-    ///
-    /// These are the pre-widened property types shown in error messages.
-    /// The `shape_id` is the widened (interned) shape; `props` contains
-    /// the original literal types from the source code.
-    pub fn store_display_properties(&self, type_id: TypeId, props: Vec<PropertyInfo>) {
-        self.display_properties.insert(type_id, Arc::new(props));
-    }
-
-    /// Retrieve display-only properties for a fresh object literal.
-    ///
-    /// Returns `None` if no display properties were stored (i.e., the
-    /// object type was not a fresh literal or had no widened properties).
-    pub fn get_display_properties(&self, type_id: TypeId) -> Option<Arc<Vec<PropertyInfo>>> {
-        self.display_properties.get(&type_id).map(|r| r.clone())
-    }
-
-    /// Store a reverse mapping from an evaluated Application result to its
-    /// original Application TypeId for diagnostic display.
-    ///
-    /// Called during Application evaluation so that the formatter can recover
-    /// the named form (e.g., `Dictionary<string>`) when it encounters the
-    /// evaluated type (e.g., `{ [index: string]: string }`).
-    pub fn store_display_alias(&self, evaluated: TypeId, application: TypeId) {
-        // Only store if the evaluated type differs from the application
-        // (i.e., evaluation actually produced a different type).
-        if evaluated == application {
-            return;
-        }
-        // A generic alias can evaluate to a bare type parameter (for example
-        // `type Id<T> = T` or a conditional alias branch that returns `T`).
-        // Type parameter ids are scoped semantic identities, not fresh display
-        // surfaces for one alias application. Recording `T -> Alias<...>` here
-        // repaints unrelated uses of `T` in later diagnostics.
-        if matches!(self.lookup(evaluated), Some(TypeData::TypeParameter(_))) {
-            return;
-        }
-        // Only alias types produced by this evaluation. Generic helper aliases
-        // can otherwise repaint unrelated earlier structural types that happen
-        // to intern to the same shape. Concrete applications remain eligible so
-        // named library/interface instantiations can display their nominal form.
-        let application_is_alias =
-            matches!(self.lookup(application), Some(TypeData::Application(_)));
-        let application_has_generic_args = application_is_alias
-            && self
-                .lookup(application)
-                .and_then(|data| match data {
-                    TypeData::Application(app_id) => Some(self.type_application(app_id)),
-                    _ => None,
-                })
-                .is_some_and(|app| {
-                    app.args.iter().any(|&arg| {
-                        crate::type_queries::contains_generic_type_parameters_db(self, arg)
-                    })
-                });
-        let evaluated_is_mapped = matches!(self.lookup(evaluated), Some(TypeData::Mapped(_)));
-        let evaluated_precedes_application = match (
-            self.lookup_alloc_order(evaluated),
-            self.lookup_alloc_order(application),
-        ) {
-            (Some(evaluated_order), Some(application_order)) => {
-                evaluated_order <= application_order
-            }
-            _ => evaluated.0 <= application.0,
-        };
-        if application_is_alias
-            && application_has_generic_args
-            && evaluated_precedes_application
-            && !evaluated_is_mapped
-        {
-            let existing_is_application =
-                self.display_alias.get(&evaluated).is_some_and(|existing| {
-                    matches!(self.lookup(*existing), Some(TypeData::Application(_)))
-                });
-            if !existing_is_application {
-                return;
-            }
-        }
-        // Never alias intrinsic types (string, number, any, etc.) — they are
-        // shared sentinels and aliasing them would make ALL occurrences display
-        // as whatever alias happened to be stored last.
-        if evaluated.is_intrinsic() {
-            return;
-        }
-        // Guard against self-referential cycles: if the Application's args
-        // contain the evaluated type itself, storing this alias would create
-        // a formatting cycle (e.g., `Wrap<T> = T | T[]` where evaluating
-        // `Wrap<{x?: "ok"}>` produces a union, and a later re-application
-        // creates `Wrap<union>` whose arg IS the union). Skip storage in
-        // that case to prevent infinite `Wrap<Wrap<Wrap<...>>>` in diagnostics.
-        if let Some(TypeData::Application(app_id)) = self.lookup(application) {
-            let app = self.type_application(app_id);
-            if app.args.contains(&evaluated) {
-                return;
-            }
-        }
-        if application_is_alias
-            && let Some(existing) = self.display_alias.get(&evaluated).map(|alias| *alias)
-            && !matches!(self.lookup(existing), Some(TypeData::Application(_)))
-        {
-            return;
-        }
-        self.display_alias.insert(evaluated, application);
-    }
-
-    /// Prefer a concrete Application display alias over structural provenance
-    /// recorded while evaluating the alias body.
-    pub fn store_display_alias_preferring_application(
-        &self,
-        evaluated: TypeId,
-        application: TypeId,
-    ) {
-        self.store_display_alias(evaluated, application);
-        if self.get_display_alias(evaluated) == Some(application) {
-            return;
-        }
-        if evaluated == application || evaluated.is_intrinsic() {
-            return;
-        }
-        if matches!(self.lookup(evaluated), Some(TypeData::TypeParameter(_))) {
-            return;
-        }
-        let Some(TypeData::Application(app_id)) = self.lookup(application) else {
-            return;
-        };
-        let app = self.type_application(app_id);
-        if app.args.contains(&evaluated) {
-            return;
-        }
-        let preserves_conditional_branch_alias = self.is_conditional_alias_base(app.base)
-            && self.get_display_alias(evaluated).is_some_and(|existing| {
-                matches!(self.lookup(existing), Some(TypeData::Intersection(_)))
-            });
-        if preserves_conditional_branch_alias {
-            return;
-        }
-        let application_has_generic_args = app
-            .args
-            .iter()
-            .any(|&arg| crate::type_queries::contains_generic_type_parameters_db(self, arg));
-        let evaluated_precedes_application = match (
-            self.lookup_alloc_order(evaluated),
-            self.lookup_alloc_order(application),
-        ) {
-            (Some(evaluated_order), Some(application_order)) => {
-                evaluated_order <= application_order
-            }
-            _ => evaluated.0 <= application.0,
-        };
-        let evaluated_is_mapped = matches!(self.lookup(evaluated), Some(TypeData::Mapped(_)));
-        if application_has_generic_args && evaluated_precedes_application && !evaluated_is_mapped {
-            return;
-        }
-        if evaluated_precedes_application
-            && !evaluated_is_mapped
-            && self.get_display_alias(evaluated).is_some_and(|existing| {
-                matches!(self.lookup(existing), Some(TypeData::Application(_)))
-            })
-        {
-            return;
-        }
-        self.display_alias.insert(evaluated, application);
-    }
-
-    /// Look up the original Application TypeId for a type that was produced
-    /// by evaluating an Application.
-    ///
-    /// Returns `None` if this type was not produced from an Application evaluation.
-    pub fn get_display_alias(&self, type_id: TypeId) -> Option<TypeId> {
-        self.display_alias.get(&type_id).map(|r| *r)
-    }
-
-    /// Record that an application base belongs to a type alias whose body is a
-    /// conditional type. This is diagnostic-only provenance.
-    pub fn mark_conditional_alias_base(&self, base: TypeId) {
-        self.conditional_alias_bases.insert(base, ());
-    }
-
-    pub fn is_conditional_alias_base(&self, base: TypeId) -> bool {
-        self.conditional_alias_bases.contains_key(&base)
-    }
-
-    /// Record the as-written origin members for a flattened Union TypeId.
-    ///
-    /// Mirrors tsc's `UnionType.origin` mechanism: when `T | null` is built and
-    /// `T` is itself a union alias, normalization flattens the result, but the
-    /// printer needs the unflattened input list to display `T | null` instead
-    /// of the structural expansion.
-    ///
-    /// We store the origin when normalization changed the visible order:
-    /// - Flattening occurred (resulting Union has strictly more members), OR
-    /// - The Union contains anonymous Object members whose canonical sort
-    ///   (by `ShapeId`) doesn't match tsc's display order. tsc displays
-    ///   anonymous objects in source/declaration order, but our interner
-    ///   sorts by `ShapeId` (allocation order), which can reorder e.g.
-    ///   `{} | { a: number }` to `{ a: number; } | {}` when the empty
-    ///   shape was interned later than `{ a: number }`.
-    ///
-    /// We DO NOT store origin for canonical sort that matches tsc for non-
-    /// anonymous-object cases (e.g., user wrote `"foo" | Refrigerator` but
-    /// the interner sorted to `Refrigerator | "foo"` — tsc does the same).
-    pub fn store_union_origin(&self, union_type_id: TypeId, origin_members: Vec<TypeId>) {
-        if origin_members.len() < 2 {
-            return;
-        }
-        let Some(TypeData::Union(list_id)) = self.lookup(union_type_id) else {
-            // Only store origins for actual Union TypeIds; other shapes have
-            // their own display paths.
-            return;
-        };
-        let current = self.type_list(list_id);
-        let flattened = current.len() > origin_members.len();
-        if !flattened {
-            // No flattening — only store if the canonical sort reordered
-            // members whose source order tsc preserves verbatim.
-            let needs_origin = self.union_origin_overrides_canonical_anon_object_sort(
-                current.as_ref(),
-                &origin_members,
-            ) || self.union_origin_overrides_canonical_number_literal_sort(
-                current.as_ref(),
-                &origin_members,
-            ) || self
-                .union_origin_overrides_canonical_keyof_sort(current.as_ref(), &origin_members)
-                || self.union_origin_overrides_canonical_application_sort(
-                    current.as_ref(),
-                    &origin_members,
-                )
-                || self.union_origin_overrides_canonical_array_pair_sort(
-                    current.as_ref(),
-                    &origin_members,
-                )
-                || self.union_origin_overrides_canonical_keyof_literal_sort(
-                    current.as_ref(),
-                    &origin_members,
-                )
-                || self.union_origin_overrides_canonical_generic_display_sort(
-                    current.as_ref(),
-                    &origin_members,
-                );
-            if !needs_origin {
-                return;
-            }
-        }
-        // First writer wins so deterministic display order is preserved when
-        // the same flattened union is reached from multiple annotation sites.
-        self.display_union_origin
-            .entry(union_type_id)
-            .or_insert_with(|| Arc::new(origin_members));
-    }
-
-    /// Replace the display origin for a union whose diagnostic context has a
-    /// more specific tsc-compatible member order than the source union.
-    pub fn replace_union_origin_for_display(
-        &self,
-        union_type_id: TypeId,
-        origin_members: Vec<TypeId>,
-    ) {
-        if origin_members.len() < 2 {
-            return;
-        }
-        let Some(TypeData::Union(_)) = self.lookup(union_type_id) else {
-            return;
-        };
-        self.display_union_origin
-            .insert(union_type_id, Arc::new(origin_members));
-    }
-
-    /// Decide whether storing the as-written origin is needed even when no
-    /// flattening occurred — i.e. the union contains anonymous Object members
-    /// and our canonical sort reordered them relative to the input.
-    ///
-    /// tsc displays anonymous (symbol-less) object union members in source/
-    /// declaration order. Our interner sorts by `ShapeId` (allocation order),
-    /// which can reorder them when the same shape is reached from earlier
-    /// annotations in the file. Returns true only when (a) at least one
-    /// member of the resulting union is an anonymous Object/ObjectWithIndex
-    /// and (b) the resulting member order differs from the input.
-    fn union_origin_overrides_canonical_anon_object_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != origin.len() {
-            return false;
-        }
-        let mut has_anon_object = false;
-        for &id in current {
-            if let Some(TypeData::Object(shape_id) | TypeData::ObjectWithIndex(shape_id)) =
-                self.lookup(id)
-            {
-                let shape = self.object_shape(shape_id);
-                if shape.symbol.is_none() {
-                    has_anon_object = true;
-                    break;
-                }
-            }
-        }
-        if !has_anon_object {
-            return false;
-        }
-        current != origin
-    }
-
-    /// Decide whether storing the as-written origin is needed for a union
-    /// whose members are all number literals.
-    ///
-    /// The canonical comparator only special-cases the `0` literal and falls
-    /// back to allocation order for other number literals. Allocation order
-    /// is global and depends on which literals tsz happens to intern earlier
-    /// (e.g., from lib processing or unrelated code), so the canonical sort
-    /// can reorder source-written `0 | 1 | 2` into `0 | 2 | 1` even though
-    /// tsc preserves the as-written order. When the canonical sort changed
-    /// the order of a literal-only union, persist the origin so the printer
-    /// can render the source-written form.
-    fn union_origin_overrides_canonical_number_literal_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != origin.len() || current == origin {
-            return false;
-        }
-        current.iter().all(|&id| {
-            matches!(
-                self.lookup(id),
-                Some(TypeData::Literal(LiteralValue::Number(_)))
-            )
-        })
-    }
-
-    /// Decide whether storing the as-written origin is needed for a union that
-    /// contains a `keyof` member. TypeScript preserves source order for
-    /// displays such as `keyof Shape | "knownLiteralKey"`, while the semantic
-    /// union comparator can place the literal first.
-    fn union_origin_overrides_canonical_keyof_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != origin.len() || current == origin {
-            return false;
-        }
-        current
-            .iter()
-            .chain(origin.iter())
-            .any(|&id| matches!(self.lookup(id), Some(TypeData::KeyOf(_))))
-    }
-
-    /// Decide whether storing origin is needed for a generated union of
-    /// same-base generic applications.
-    ///
-    /// Distributive conditional alias display builds unions such as
-    /// `ChannelOfType<T, TextChannel> | ChannelOfType<T, EmailChannel>`.
-    /// The semantic union comparator sorts same-base applications by their
-    /// type arguments, which can invert the branch order that tsc displays from
-    /// the source union being distributed. Preserve that origin only when the
-    /// canonical union contains exactly the same same-base applications.
-    fn union_origin_overrides_canonical_application_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != origin.len() || current == origin {
-            return false;
-        }
-
-        let mut current_sorted = current.to_vec();
-        let mut origin_sorted = origin.to_vec();
-        current_sorted.sort_unstable_by_key(|id| id.0);
-        origin_sorted.sort_unstable_by_key(|id| id.0);
-        if current_sorted != origin_sorted {
-            return false;
-        }
-
-        fn same_application_base(
-            interner: &TypeInterner,
-            ids: &[TypeId],
-            expected_base: &mut Option<TypeId>,
-        ) -> bool {
-            for &id in ids {
-                let Some(TypeData::Application(app_id)) = interner.lookup(id) else {
-                    return false;
-                };
-                let app = interner.type_application(app_id);
-                match expected_base {
-                    Some(base) if *base != app.base => return false,
-                    Some(_) => {}
-                    None => *expected_base = Some(app.base),
-                }
-            }
-            true
-        }
-
-        let mut expected_base = None;
-        same_application_base(self, origin, &mut expected_base)
-            && same_application_base(self, current, &mut expected_base)
-            && expected_base.is_some()
-    }
-
-    fn union_origin_overrides_canonical_array_pair_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != 2 || origin.len() != 2 || current == origin {
-            return false;
-        }
-
-        let mut current_sorted = current.to_vec();
-        let mut origin_sorted = origin.to_vec();
-        current_sorted.sort_unstable_by_key(|id| id.0);
-        origin_sorted.sort_unstable_by_key(|id| id.0);
-        if current_sorted != origin_sorted {
-            return false;
-        }
-
-        fn is_array_of(interner: &TypeInterner, array: TypeId, element: TypeId) -> bool {
-            matches!(interner.lookup(array), Some(TypeData::Array(inner)) if inner == element)
-        }
-
-        is_array_of(self, origin[0], origin[1]) || is_array_of(self, origin[1], origin[0])
-    }
-
-    fn union_origin_overrides_canonical_keyof_literal_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != 2 || origin.len() != 2 || current == origin {
-            return false;
-        }
-
-        let mut current_sorted = current.to_vec();
-        let mut origin_sorted = origin.to_vec();
-        current_sorted.sort_unstable_by_key(|id| id.0);
-        origin_sorted.sort_unstable_by_key(|id| id.0);
-        if current_sorted != origin_sorted {
-            return false;
-        }
-
-        let is_keyof = |id| matches!(self.lookup(id), Some(TypeData::KeyOf(_)));
-        let is_literal = |id| matches!(self.lookup(id), Some(TypeData::Literal(_)));
-        (is_keyof(origin[0]) && is_literal(origin[1]))
-            || (is_literal(origin[0]) && is_keyof(origin[1]))
-    }
-
-    /// Preserve source-written order for generic display unions whose canonical
-    /// member sort is not the order tsc prints in diagnostics.
-    ///
-    /// Pure-literal unions are excluded: when every member is a literal,
-    /// our canonical alloc-order sort already matches tsc's
-    /// registration-order display (e.g. for instantiation-time unions like
-    /// `T | U` with `T="World", U="Hello"` where `"Hello"` was registered
-    /// first, tsc displays `"Hello" | "World"`).
-    fn union_origin_overrides_canonical_generic_display_sort(
-        &self,
-        current: &[TypeId],
-        origin: &[TypeId],
-    ) -> bool {
-        if current.len() != origin.len() || current == origin {
-            return false;
-        }
-
-        let mut current_sorted = current.to_vec();
-        let mut origin_sorted = origin.to_vec();
-        current_sorted.sort_unstable_by_key(|id| id.0);
-        origin_sorted.sort_unstable_by_key(|id| id.0);
-        if current_sorted != origin_sorted {
-            return false;
-        }
-
-        let has_complex_generic = origin.iter().any(|&id| {
-            matches!(
-                self.lookup(id),
-                Some(TypeData::Application(_) | TypeData::KeyOf(_) | TypeData::IndexAccess(_, _))
-            )
-        });
-        if has_complex_generic {
-            return true;
-        }
-
-        // Mixed unions where literals coexist with non-literal complex types
-        // still need origin preservation, but pure-literal unions don't.
-        let all_literals = origin
-            .iter()
-            .all(|&id| matches!(self.lookup(id), Some(TypeData::Literal(_))));
-        if all_literals {
-            return false;
-        }
-
-        origin
-            .iter()
-            .any(|&id| matches!(self.lookup(id), Some(TypeData::Literal(_))))
-    }
-
-    /// Look up the as-written origin members for a flattened Union TypeId.
-    pub fn get_union_origin(&self, type_id: TypeId) -> Option<Arc<Vec<TypeId>>> {
-        self.display_union_origin.get(&type_id).map(|r| r.clone())
     }
 
     pub(super) fn intern_function_shape(&self, shape: FunctionShape) -> FunctionShapeId {
